Introducing compatibility with FerretDB

+++ components/storage/test/migrations/1.7.5.test.js
@@ -54,10 +54,12 @@ describe('Migration - 1.7.5', function () {
   let accessesCollection;
 
   before(async function () {
+    if (database.isFerret) this.skip();
     accessesCollection = await database.getCollection({ name: 'accesses' });
   });
 
   after(async function () {
+    if (database.isFerret) return;
     // erase all
     await accessesCollection.deleteMany({});
   });
